Fix kernel in cobbler and add Yarmouth PPA
The idea behind `YARMOUTH_PPA_USERNAME` and `YARMOUTH_PPA_PASSWORD` is
to be passed via make variable (or environment variable) to the build
script so that it's not given directly on the command line. The `make
build` will automatically grep all the sub-projects in search for the
`YARMOUTH_PPA_USERNAME` phrase and replace occurences of both phrases
with the value of a respective make/environment variable.

JIRA:FUEL-39