Updated armband Project: fuel  ba99c6e97a4dc9db8165edc7d95a254892f59830
Build: Allow customizing FUEL_MAIN_TAG (commit)

FUEL_MAIN_TAG is currently hard set to 9.0 (not overridable).

While this is a sane assumption for the usual scenarios,
Armband applies a series of patches on top of Fuel@OPNFV, hence
requiring us to override this variable.

WARNING: FUEL_MAIN_TAG is reused for fuel-mirror tag, so if you
override this variable, make sure the same tag exists in fuel-mirror.

Change-Id: Ided75cf0c3b5ad18cf7ef1ec88b5d2dc3aada511
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>
Signed-off-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>