Updated armband Project: fuel  e7a78baae49f02b4303b6204eea7de01c527744c
Enable merge of the versions struct of a plugin

Redoing this the reverted patch. The original idea
that everyone which wants to merge lists has to
code that, is not wanted. So the Exception will
become a warning.

This reverts commit 552536f5319d6ead73118d0cfd701d648e99df28.


(cherry picked from commit fb238de7c9833e91771ee4448f425f45168f8e6e)