target Ubuntu: Blacklist rtc-efi if not supported.
Older ThunderX and possibly other UEFI-enabled targets do not support
rtc-efi properly, so they end up filling dmesg with useless
complaints about not being able to read system time.

This adds a simple test for rtc-efi and blacklists it in target OS
(Ubuntu) if not supported.

While we're at it, rename a patch file I missed in a previous rebase
and update TODO.