build.sh: do not use readlink for unexisting dir
The build failed to copy the artifact files to the destination
directory because the variable BUILD_DIRECTORY was empty.
This variable is obtained doing "readlink" on a path, although
at the time of the assignment that path does not exist yet. This
is because it depends on a path inside a submodule that has not
yet been initialized.

We remove readlink on the assignment since we know the path will
already be in a canonicalized form, since it is constructed from
an already canonicalized path.


Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>