build.sh: do not use readlink for unexisting dir
the fuel submodule is empty until make submodules-init is done, which
happens after the initialization of the BUILD_BASE variable, which
does "readlink" on the path, but since the path doesn't exist yet,
the variable is left empty, making the build fail.

We remove readlink on the path since we know the path will already be
canonicalized, because it is constructed from an already canonicalized
path.

Change-Id: Icc1113d26b503ff998aebbcd1efe81004859f878
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>