Updated armband Project: fuel  3b8db706acb60b680bcf220bd0236d5efcc66d07
common.py: catch stderr in exec_cmd

When running commands with exec_cmd(), the stderr of the command is
sent to /dev/null and ignored, and only stdout is retrieved. Thus, when
a command fails and check is enabled, only the output of stdout is
presented to the user, which normally holds no information about the
error.
In this patch we retrieve stderr, and when an error occurs, an exception
is raised with that message.

Fixes https://jira.opnfv.org/browse/FUEL-142

Change-Id: I3940e1a43963a6abec362481b1d4ce7bd7cb816d
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>