Patch set for Armband deploy scrips
These are a collection of patches that adapt the current Fuel deploy
scripts to do several things:
  - Make it possible to create a Fuel VM on a remote libvirt server.
    We use the LIBVIRT_DEFAULT_URI environment variable to detect that.
    For local deploys it is possible by setting it to "qemu:///system",
    or leave it empty.
  - Make it possible to add additional network interfaces. For this we
    allow the user to pass the "-b bridge" paramter several times, and
    creating a new virtual NIC for each of them, in the same order they
    were given.
    This required a bit of refactoring of the code.

None of the changes above should break backwards compatibility.

In addition there are some updates to the code that were deemed
necessary, like the ability to retry when executing shell commands,
instead of directly failing, and a simplification of the DHA IPMI
adapter.

Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>