[fuel] deploy.sh: Add timeout flag on fuel
Some PODs might need to increase the default timeout value.
The patch introduces the -T flag to Fuel's deploy script, that
can be used to set the timeout. If the flag is not given, it
will look into the DEPLOY_TIMEOUT environment variable and use
that value instead, otherwise the default vaule in deploy.py
will be used.

Change-Id: I835473190673dd8646d0da48c2dbd3bb5e13046d
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>