Increase maximum shell command timeout to 2h
Since `execute_shell_command` mcagent is used for building the target
image and the timeout provided in the astute task is just respected on
the Astute level (on mcollective level this timeout is actually a
minumum between value provided and ddl value) we need to increase it
for Armband operation.

Because Armband builds the target image via `qemu-debootstrap`, it can
take more than 1 hour to finish the building. Therefore aside of
increasing the timeout in the Astute task, the maximum timeout value
for mcagent has to be increased.

Change-Id: Iab94e794e36bb7e887e8f334daa46e82fe0c6f94
Signed-off-by: Stanislaw Kardach <stanislaw.kardach@caviumnetworks.com>