docs/install-instr: Refactor target specific cfg.
Fuel 8.0 adds support for setting up "Offloading Modes" for
target nodes.

Document how this features should be used (and why/when it's
necessary).

While at it, fix minor reference to branch in git tag checkout
instructions, as we'd be in detached HEAD and not on a new branch.

Change-Id: If7a8ca44541903dbeb6db36bc64fe56b132165cb
(cherry picked from commit 3cb6402c59bbf72099e63fd3f26017089520841e)