Increase open file limit for OpenDaylight
Upstart scripts by default do not run scripts in a shell which results
in limits from `/etc/security/limits.conf` not be taken into account
and results in the default value (4096) to be taken. This may prove
problematic for OpenDaylight during initialization. On Cavium ThunderX
we have noticed OpenFlow controller initialization failures caused by
hitting this limit.
OPNFV is increasing the `nofile` limit to 112640 for OpenStack services
which is why I'm using this value here.
Bug report filed to Fuel@OPNFV and will be released in Colorado (see
[1]).

This is a cherry-pick of 39f77acb6d4cdcbea04a817b01f67cad8f202c1d since
the fix in `openstack/fuel-plugin-opendaylight` targets C-release and
this bugfix is critical for arm64.

[1] https://jira.opnfv.org/browse/FUEL-140

Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>