Optimize lshw invocation
nailgun-agent triggers lshw periodically and lshw takes too much time
on some hardware like Cavium ThunderX crb2s (close to 5 min) but this
hardware information is constant unless you add new hardware into
nodes.

This patch optimizes subsequent runs of nailgun-agent by saving the
lshw output to a file for the first time it is triggered and reading
it from that file on consecutive runs.

lshw output file is saved on the tmp filesystem and on reboot it will
be cleared and therefore any hardware changes will be detected.

Signed-off-by: Vijayendra Radhakrishna <vradhakrishna@mvista.com>
Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>