Rebase: p/fuel-web: Drop obsoleted qemu doc patch.
Upstream moved documentation from local repo to fuel-docs in [1].

While looking at this patch, I think we should drop this change
altogether, since qemu-user-static is not needed inside the
docker builder container, at least not until we start using it
for cross-builds, which is highly unlikely anyway.

qemu-user-static should always be automatically installed as a
dependency of fuel-agent, and NOT by explicit request.

[1] https://github.com/openstack/fuel-web/commit/
    eb673991cc99c5dadb251f922adb08abb40a8bad

Change-Id: Iad7f659fd5ff81e92860508fad58bac519158c47