deploy: ipmi adapter: Port config support.
Sometimes the IPMI lanplus protocol listens on a non-standard
remote port, e.g. when target nodes are interfaced through a
fake IPMI BMC application that listens on multiple ports on the
same IP address.

Therefore, allow setting IPMI port in DEA/DHA and pass it along
to `ipmitool` when set.

Change-Id: Ibff1309b17733af1a6fc8c5008495f42d7ef084a