FIXME: Temporary: Remove repo-mirror-cz.patch.
Upstream removed all mirrors but usa and cz, and changed
its default to mirror.fuel-infra.org in [1].

This allows us to drop patching of fuel-main/config.mk
with the Czech mirror.

If needed, USE_MIRROR=cz can be specified as an env var.

FIXME: Remove when Fuel@OPNFV pulls change and deals with it.

[1] https://github.com/openstack/fuel-main/commit/
    751d502cfe15d9c9df0ee89530ac3b0b73aa1638