Fix: nova patching: broken puppet chain.
This was an interesting bug. The code has been broken for a while,
but we did not notice because nova was installed before these
puppets were ran.

With the Mitaka reworks and changes in the puppet manifests, this
bug surfaced, but manifested as `virtlogd` service not being
started, looking like a libvirt 1.3+ problem.

So, just add the missing "->" in the troubling puppet file.

Change-Id: Ic702e17e8f732bf7a277714cc4416e0f2ee68cfe
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>