Update patch set 2
Patch Set 2: Code-Review-2

This is changing the behavior of the upstream fuel@opnfv. The scenarios are meant to be applied on top of the DEA, which is where the roles of the different nodes are defined. What we will do with this patch is install always the same scenario: the one defined in our dea override...

The issue is deeper than I thought initially. Indeed the current scenarios do not contemplate the possibility of PODs with different hardware that require different transformations_ and interfaces_...

Patch-set: 2
Reviewer: Gerrit User 486 <486@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-2