iptables: Move SSH rules to post-scripts.
Previous change [1] introduced the addition of an iptables rule
that was supposed to allow SSH access on all ifaces (not only
admin iface) when additional ifaces are configured.

However, Fuel installer is flushing the rules after transplant
adds our SSH config, overwriting it.

Move iptables SSH config to post-install section, as standalone
script. In order to keep the same behavior, test ifcfg-eth0
(admin interface is expected to be called eth0 by convention)
for "DEFROUTE=no" and only whitelist SSH on all ifaces if it matches.

[1] https://gerrit.opnfv.org/gerrit/#/c/16571/

Change-Id: I086b75461daa62671cad10494fe34acfd77757ae
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>