Updated armband Project: fuel  3ec3fb2d1d316b656a13f9c1ca00cc8a7fcd51de
deploy-config: honor interfaces and transformations

Currently all scenarios assume interfaces and transformations are the
same for all nodes in the POD, however some PODs may contain nodes that
have different hardware, or where the interfaces are configured
differently.

In this patch we honor the original interfaces and transformations if
they are present in the dea-override.yaml file. The way to add this
information in the dea-override is by having a "nodes:" section with
this information, ie:

nodes:
- id: 1
  interfaces: interfaces_1
  transformations: transformations_1
- id: 2
  interfaces: interfaces_2
  transformations: transformations_2
- id: 3
  interfaces: interfaces_1
  transformations: transformations_1

The node IDs is used to find out this information.

Change-Id: If6ff8ca28b42e043d1bdf91142a4a56ae36e4304
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>