fuel-astute: Erase EFI boot entry on EFI systems
On EFI-enabled systemd, grub-install from grub-efi-* package
installs a boot entry named "ubuntu".

Since this boot entry is saved in board flash memory, erasing
the MBR bootloader code and/or partition signature will not
clear the ubuntu boot entry, leaving it pointing to a loader
on the ESP (EFI System Partition) that will try to load the
kernel/initrd from an erased partition.

In Fuel 8.0, the whole disk was erased, so the ubuntu EFI boot
entry was skipped due to missing ESP, while for Fuel 9.0 and
above we have to explicitly remove it to keep the system in a
sane state.

Note: efibootmgr is installed automatically on EFI systems
as a dependency of grub-efi-*.

Closes: ARMBAND-47

Change-Id: If715e960dfd96df807abb0e26f4b7604624ac408
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>