Makefile: Trim OPNFV_PRODNO to fit in ISO volid
genisoimage has a limit of 32 chars for volume ID, which was
reached during Jenkins builds (which sets REVSTATE to a datestring),
but not while building outside Jenkins (in which case REVSTATE was
simply set to "P000"), hence the recent broken builds despite
manully testing the changes beforehand.

Shorten the product name (PRODNO var), and revert git SHA to full
format, as that was not the culprit.

Related-bug: ARMBAND-53

Change-Id: I0a530982893af1b6a0243b55f6b1d2eea66ac5bf
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>