Updated armband Project: fuel  6222031ed4e195fee85d6e3bdb5f34845fa061e9
repo mirror: Allow multi-arch local mirrors

Original patch from Stan, signed by me for upstreaming.

This patch allows specifying multiple architectures via UBUNTU_ARCH in
form of a list of space separated architectures. The first architecture
in the list is considered primary and will be used for building all the
deb packages by fuel-main. Additional architectures are added to allow
targets of other architectures to use the mirror.

NOTE: this imposes a requirement that all packages built are arch
independent (which is true so far).

[ Alexandru Avadanii ]
Reworked for applying on top of Fuel@OPNFV.
Pass UBUNTU_ARCH to Docker containers.

Change-Id: Ibc4ae03223545383813ac58ec34d44c1f97b4694
Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>

build: cache: Consider UBUNTU_ARCH in .cacheid

Usually UBUNTU_ARCH is not changed inside the same repository,
but we should consider this case anyway for future merging
Armband and Fuel@OPNFV repositories/build systems.

This applies to all Fuel plugins and the local repo.

Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>