Updated armband Project: fuel  2e6be7473d11f14d62ad1aa1d76f2e0c99327424
check read access instead of isfile() for Fuel ISO

When testing with nested virtualzation I attached
the Fuel ISO to VM (/dev/sr0) running deploy.sh

This failed because we expect the ISO to be a file, when
it could be a device, e.g. /dev/sr0.  Instead of checking isfile()
check for read access.

This tested with bsdtar. It can extract an ISO from a file or device.

Change-Id: Ic09373cd618b33382303d9134cce4e005040de95
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>