Add a virtual pod for ARM
This is the first virtual pod added for ARM. The current fuel arhitecture
assumes that all the VMs (Jenkins slave, Fuel master and Target nodes) are
all on the same server. There is code that allows the Fuel master to be on
a different machine, but nothing for the Target nodes.
In Enea-Arm lab we have a different server for each.

Therefore, the following changes are done:
- The default libvirt URI is the machine on which the Target nodes are
- The parameter libvirtUri was added in the dha.yaml file for fuel master
- All virsh calls have "-c" option to explicitly specify the uri; this parameter
  is set only when libvirtUri is explicitly set in the yaml file
- The scenarios dha override now point to the arm template files
- The network templates are not needed for ARM, as the network infra is provided
  by the server on which the Target node VMs reside; the folder is added here
  just for consistency
- The target nodes are created with virsh rather than quemu when the nodes are
  not on the same machine as Jenkins slave (similar logic to what is done for
  remote Fuel master)

jira:ARMBAND-122

Change-Id: Ibf4744e5593f5201550115397024fd8311696074
Signed-off-by: cristinapauna <cristina.pauna@enea.com>