Correct previous patch on fuel rework
Some leftover chars were keeping the deploy to work
Also, the empty folder networks was added. This folder should
contain templates for virtual network config for the targets.
For arm we don't need those so we leave it empty for consistency
with other projects and for avoiding changing the deploy logic

Change-Id: I9ce00157e5cddcd14b8225fb04ea057f45461406
Signed-off-by: cristinapauna <cristina.pauna@enea.com>