Correct previous patch on fuel rework
Some leftover chars in https://gerrit.opnfv.org/gerrit/#/c/24901/
were keeping the deploy to work.
Also, the empty folder networks was added. This folder should contain templates
for virtual network config for the targets. For arm we don't need those so we
leave it empty for consistency with other projects and for avoiding changing
the deploy logic

JIRA: ARMBAND-122

Change-Id: Id33e2efacc3ff8d3754c8b29d4bebefd0b385c41
Signed-off-by: cristinapauna <cristina.pauna@enea.com>