fuel-plugin-opendaylight: Upstart service: Respawn
On arm64, JVM sometimes (~10%) crashes at the very start, due to
JIT issues in our openjdk8 package.

We can't really solve the JVM on arm64 issues in the near future,
so we will try to work around the issue by configuring the ODL
service as respawn-able (see [1] for more details).

Note: This is specific to Ubuntu Trusty (14.04), which uses upstart,
for Ubuntu Xenial (or anything else >14.x), which uses systemd,
a different, equivalent mechanism should be used.

[1] http://upstart.ubuntu.com/cookbook/#respawn

JIRA: ARMBAND-134

Change-Id: I0dc82e4a6d0cc485e981477e6a25dbad95db930b
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>