Fix failed deploys on virtual pod (noha scenarios)
JIRA: ARMBAND-173
Currently only ha scenarios were successfully deployed on virtual pods
The default configuration is for ha scenarios and those have 5 target nodes
while on the noha scenarios there are 4. The noha scenarios failed because
it was trying to parse one extra node.

This patch modifies the dea and dha files for each noha scenario so that the
nodes are properly configured.

Change-Id: I360aeaf8dc92e0ff614367d47d74f6fa61ef2268
Signed-off-by: cristinapauna <cristina.pauna@enea.com>