fuel-library: pp: cobbler: Fix CentOS ver detect
Upstream fuel-library bug report: [1].

Recently, CentOS released version 7.3.1611, which triggered a latent
bug in fuel-library puppet scripts related to differentiating between
6.x and 7.x major versions: 7.3.1611 is wrongly detected as 6.x,
leading to the wrong packages being attempted to be installed,
leaving nailgun server service broken on the Fuel Master node.

This change forces the match at the beggining of the version string
(e.g. "/^6.+/: {" instead of "/6.+/: {"), so the new CentOS 7.3.1611
no longer matches the condition for 6.x).

This patch will soon be obsoleted by the Netwon switch.
However, until that happens, we need to ensure CentOS 7.3 is not
breaking the deploy.

[1] https://bugs.launchpad.net/fuel/+bug/1649895

JIRA: FUEL-238

Change-Id: Iebda7e13d66191614c3552dacb6c80091215baf5
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>