submodules: Disable ignore=all for OPNFV projects
Armband has a number of git submodules, which are hosted:
- via Openstack (review.openstack.org, github.com);
- via OPNFV (gerrit.opnfv.org);

We cannot automatically detect changes in Openstack submodules,
but within OPNFV, an "Updated submodule" commit is generated
when a change is merged into the submodule git, provided we
do not have an "ignore=all" git config directive.

Ignoring submodule changes also prevents automatic ISO builds in
Armband when Fuel@OPNFV changes are merged, which is something we
needed during the D-release development cycle.

The downside of removing "ignore" is that submodules are now
seen as "modified" by git after apply our Armband patches.

While at it, remove one fuel-library patch that got upstream, just
to force a new ISO build.

Change-Id: I8e81e781cb5c78c920a2a33cd8d584de3d9d4fd7
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>