fuel-plugin-odl: Fix deploy on x64 due to leveldb
leveldbjni class is used to patch odl for aarch64
When run on x64, the class is not declared (it is filtered out).
Nevertheless, the puppet dependecy required the now undeclared class.

Fix consists in always using leveldbjni class, but on x64 it does
nothing. The architecture dependent logic is thus contained in leveldbjni.

JIRA: ARMBAND-229

Change-Id: If61f4fe3dcca0b2cff38ffcfc2321ed4b43bc2b4
Signed-off-by: Dan Andresan <dan.andresan@enea.com>