p/fuel: deploy: Allow non-root deploys
Fuel@OPNFV's ci/deploy.sh script imposes a hard requirement on
running as the root user, which can be refactored into a series of
smaller requirements.

Instead, relax the check to the ability to run sudo commands
inside the script, and explicitly add sudo for the package
installation commands.

This also adds a requirement for the Jenkins user running the
deploy to be part of the "kvm" and "libvirtd" groups, so
libvirt/virsh can manage VMs on the jump server / Jenkins slave.

See related change in OPNFV Releng repository [1].

[1] https://gerrit.opnfv.org/gerrit/#/c/37655

Change-Id: I29b90d285c99f85462493c2a3890264365fca9f0
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>