ci/deploy.sh: Do not pass SSH_KEY to Fuel deploy
OPNFV Jenkins jobs for Fuel and Armband define the variable SSH_KEY,
which gets passed down to Fuel's ci/deploy.sh, leading to a conflict
between the original key and the destination key paths, which in
the default case overlap, leading to a deploy-breaking fault.

Fuel@OPNFV is not affected by this behavior due to a "sudo" call
that does not pass down enviroment variables (i.e. without `-E`).

Change-Id: I291966ee177d828bf1e59c213c9da40eae13ab6b
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>