p/fuel: network: public: Use arm-virtual2 POD cfg
FIXME: This should be converted into a dynamic configuration read
from the universal POD descriptor in securedlab, once that is ready.

Until then, just align the public network configuration used by the
virtual POD with the Enea lab configuration specific to arm-virtual2
(i.e. public network on 10.0.9.0/24).

NOTE: Instead of replacing the gateway 10.16.0.1 (now 10.0.9.1) with
our lab's gateway (10.0.9.254), keep it as an extra hop, in order not
to break deploys in different enviroments.

Also, since our gateway resides at 10.0.9.254, trim the DHCP pool
range for the public network to not include that address.

Change-Id: I92f01dd819ebf41b1da68d3b934fb10a23257e33
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>