salt.sh: Add "arm64" to bootstrap arch condition
Upstream commit [1] broke Salt bootstrap on AArch64, by
introducing an architecture condition that is too strict to allow
Debian package installation (even if we provide our own repo).

Add "arm64" to the list of supported architectures. This needs
to be done on the fly, as the bootstrap script is fetched using
`curl` from <salt-master-setup.sh>, which is also fetched
dynamically using `svn export`.

[1] https://github.com/saltstack/salt-bootstrap/commit/caa6d7d

Change-Id: I2ff5d56e4d3072a1afe31ab0a400b0b83e2b34a8
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>