Update patch set 1
Patch Set 1:

Perhaps we can add some more comments to this patch. I assume that you want to reuse the mcp.rsa in order to change the boot entries from the older nodes. 
That is unclear however to other members. Perhaps we should elaborate more in the commit message ?
guhe ?

Patch-set: 1