Update patch set 1
Patch Set 1:

> Perhaps we can add some more comments to this patch. I assume that
 > you want to reuse the mcp.rsa in order to change the boot entries
 > from the older nodes.
 > That is unclear however to other members. Perhaps we should
 > elaborate more in the commit message ?
 > guhe ?

Hi, this is something I wanted to do before the UEFI thing landed, basically we don't need to regenerate a SSH key for each deployment. Instead, if the user wants to re-use a personal ssh key, now they can do that by copying it to a predefined location (/tmp/mcp.rsa for now, but I hope to change that to something persitent later).
I will add more comments to this patch when I propose it upstream in Fuel@OPNFV.

Patch-set: 1