Update patch set 1
Patch Set 1: Code-Review+2

This should speed up the deploy with a few minutes, since salt.control (responsible for preparing and launching VCP VMs) handles things sequentially, i.e. for each VCP VM, the base image is fetched, locally mounted, salt-minion is installed, then the VM is spawned.
When we pre-install the salt-minion, we only have to do it ONCE for the whole cluster.

Patch-set: 1
Label: Code-Review=+2