Update patch set 6
Patch Set 6: Code-Review-1

> (1 comment)
 > 
 > I think all the comments have been fixed

No, this will not work with Mirantis packages when we switch to them.
In order for your patch file to be idempotent when applied to both Ubuntu sources and MCP/upstrem sources, it has to contain the exact contents of the James' patch reverted (as in `git revert` or `patch -R`).

Patch-set: 6
Label: Code-Review=-1