Update patch set 6
Patch Set 6:

I dont understand what you mean. The patch looks exactly like it reverts the changes from that james-page commit. The only differences are the line numbers and the testcase.
The testcases we said to not revert.
As for the line number, I cannot do it like a git revert because there are other patches which apply to that file. So some numbering of lines is not the same. 
I do not know what you want me to change.

Patch-set: 6