Revert "[p/openstack.nova] force creation of pty on aarch64"
Now that we added Mirantis MCP repos on cmp nodes for HA scenarios,
the recently added patch that reverted Ubuntu specific changes for
serial console no longer applies (as Mirantis does not carry those
Ubuntu modifications but instead relies on upstream Nova).

This, combined with a minor bug in patch applying itself and a
legacy leftover that does not check the patch apply return code
led to the whole Nova patch to be skipped.

- remove pty related patch hunks (revert);
- fix `unless` condition for Nova patch apply, as well as Salt minion
  patch apply;
- break `armband` state on patch apply failure;

Note that for noHA deploys (where we still use the Ubuntu packages
on cmp nodes afair) we will revert to old (no pty) behavior.

This reverts commit 093f2b7684c86979dfa6f5f36f28689ca9b3e552.

Change-Id: Ib3c1119ab544ba895b2f8d37e0b9542196058993
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>