Update patch set 4
Patch Set 4:

Well If x86 wants to switch as well we have to put it somewhere else and not in armband. That being said x86 probably has not even noticed the error in functest because they have 1 more pci slot available.
The question is now more what to do with 0012

Patch-set: 4
Label: Verified=0